Remove UCX run directories which are now identical to standard
With the simple SOA updates, there is no longer a difference between the
Standard and UCX simulations. Therefore, we can remove the UCX run
directories and options from the GEOS-Chem unit tester.

Also moved the standard *.dat files for FAST-JX photolysis from the
tropchem directory to the standard directory in chem_inputs and added
symbolic links in the tropchem directory.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>