Move GCHP HEMCO config file into chem_inputs/standard
The HEMCO_Config.template file in gchp_4x5_standard has been moved to
chem_inputs with the rest of the HEMCO config files and a symbolic link
pointing to that file is now in the run directory. Also changed symbolic
links for the *.dat photolysis files to point to chem_inputs/standard
instead of chem_inputs/tropchem.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>