Removed advected species option from runConfig.sh
Including the total number of species and list of species in runConfig.sh
is misleading since the script does not yet allow removal of species.
The user would therefore have to both run the script with an updated list
as well as update input.geos manually. It is better to not include the
option in runConfig.sh to make it clear changes must be made in the
other file.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>