Use reprocessed NEI2011 monthly mean files that have 4th dimension removed
All NEI2011 monthly mean files recently processed from the hourly files
have a 4th dimension (nb2) incompatible with MAPL in GCHP. The 4th
dimension is used in a new variable containing the time boundaries of
the hourly data used for calculating the mean. GCHP now reads a version
of each of the files with the new time boundary information removed.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>