Update GCHP slurm run script if error encountered in runConfig.sh
Prior to this commit, run configuration problems found when
executing runConfig.sh would print the error but not result in
skipping the GEOS-Chem simulation. This would lead to hanging
runs that eventually would time out. With this update, the
environment will be loaded and the simulation run only if
runConfig.sh exits with 0. This will result in immediate
termination if a setup error is found.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>