Use soilnox fertilizer file with corrected time vector for GCHP
The new file is the same as soilNOx.fert_res.generic.05x05.nc except with
time variable vector changed from 1-366 to 0-365. The time unit is days
since Jan 1 00:00:00 and therefore the first entry should have time=0
for correct read-in by GCHP ExtData. Time=1 is interpretted as Jan 2
00:00:00 which is the incorrect date for that data slice. This causes
GCHP to exit with an error if starting a simulation on Jan 1.

This file is being used for the v11-02b GCHP 1-year benchmark. It will
replace soilNOx.fert_res.generic.05x05.nc in v11-02c.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>