Add a sample HISTORY.rc to GEOS-Chem classic rundirs
We have added a very simple HISTORY.rc file for unit testing the
new netCDF diagnostics.  This will archive only the advected
species.

The original file is runs/chem_inputs/standard/HISTORY.rc and is
linked to all of the GEOS-Chem Classic run directories.  The
HISTORY.rc file will be ignored unless you compile with NC_DIAG=y.

We did not touch the HISTORY.rc files in the gchp_c24_* rundirs,
as they are specifically customized for GCHP.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>