Restore complex SOA mechanism as default in full-chemistry simulations
The GCSC has decided that the complex SOA mechanism will be the default
and that the simple SOA scheme will be an option in v11-02c and later.
The SOAP and SOAS species have been removed from the input.geos.template
and HEMCO_Config.template files for all full-chemistry simulations and
have been replaced with the complex SOA species (MTPO, TSOA/G, ISOA/G,
ASOA/G). In addition, Eloise Marais's isoprene SOA option is now turned
on by default.

A new run directory for simpleSOA has been added for running the simple
SOA scheme with the SOAP and SOAS species. The aerosol-only simulation
also still includes the simple SOA species, but this can be updated later
if needed.

Finally, the 4x5_isop_soa run directory has been removed. It is not needed
anymore because Eloise's isoprene SOA option is turned on by default.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>