Update input.geos.template files to remove support for old met fields
In the input.geos.template files, the lines for setting the GCAP,
GEOS-4, GEOS-5, and MERRA data directories are no longer needed. The
lines for setting temporary directory, the option for unzipping met
fields, and the option to wait for met fields to be unzipped have also
been removed as they were only needed for the old BPCH met fields.

Also updated README files to remove references to obsolete met fields.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>