Update makefiles for netCDF diagnostic output
We have made the corresponding changes in the unit-test and
diff-test makefiles so that they will look for the netCDF diagnostic
output files generated by the GEOS-Chem history component. This is
specified in the HISTORY.rc file as the "inst" collection, so the
base file names will be "GEOSChem.inst.YYYYMMDD.nc4".

Also updated the perl/getRunInfo script to return just the
YYYYMMDD part of the start date, which is used in the netCDF
diagnostic file names (rather than YYYYMMDDhhmm).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>