Add run directories for halogen simulations
Run directories for 4x5_halogens and 2x25_halogens have been added.
A new directory has been created in runs/chem_inputs/ for the halogen
simulation to keep the FJX_*.dat and HEMCO_Config.template files for
that simulation separate from the files for the rest of the fullchem
simulations. This may change after the halogen simulation has been
evaluated and when it has been whether or not the halogen chemistry will
be default or not.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>