Add fix to ND51 timeseries diagnostic for benchmark simulations
The 1-year benchmark simulations utilize the ND51 timeseries diagnostic
for saving out data for BrO satellite comparisons. The input.geos file
in 4x5_benchmark/ was set up to save out BrO, 3-D cloud fractions, and
cloud top heights. However, as of v11-02c the number of advected species
exceeds 150 in the benchmark simulation, so using values 154 and 156
no longer saved out the cloud fields. A fix has been added to the code
so that all non-species quantities are indexed with 501 and above. The
values have been fixed her to assist in creation of 1-year benchmark
run directories.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>