Fix update interval in HISTORY.rc files for nested geosfp, TOMAS sims
For the nested GEOS-FP simulations at 0.25 x 0.3125, we now set
the inst.frequency ("update") interval to 001000.  This corresponds
to the chemistry timestep used in the input.geos.template for
GEOS-FP nested runs.

Similarly, for the TOMAS simulations, we now set inst.frequency to
30 minutes (as was done in the prior commit).

For simplicity, we have added two new files:

  runs/chem_inputs/standard/HISTORY.rc.geosfp_nested
  runs/chem_inputs/specialty/HISTORY.rc.TOMAS

which are symbolically linked to the appropriate GEOS-FP 0.25 x
0.3125 nested grid run directories and TOMAS run directories.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>