Add -no-requeue to the SBATCH tags atop the job script
In case a node fails and then reboots during a unit test run,
the SLURM scheduler will automatically re-submit the unit test
job.  But this will likely cause an error, because the unit test
will not be starting from a "clean" configuration.

FASRC suggested to add "#SBATCH --no-requeue" to the list of
SLURM tags at the top of the script.  This will prevent the
unit test job from being resubmitted.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>