Include UTC hour and minute in GCHP output filename template in HISTORY.rc
This update is necessary to prevent the following issue in GCHP:

File output is incomplete if outputting 1-hr time-averaged data to daily
files using filename format %y4%m2%d2.nc in HISTORY.rc. For example, if
running GCHP with 1-hr output frequency and 1day duration for a 1-day
simulation the output file should contain 24 time values. The first time
should be 30 minutes after the start time, e.g. 00:30:00 if starting the
simulation at midnight. However, if HISTORY.rc is configured to show time
format string YYYYMMDD in the filename then only one time is in the file
and it corresponds to the last of the 24 averaged taken over the day (e.g.
 23:30:00).

This problem will also appear for other combinations of file duration,
frequency, and filename timestamp string. You can avoid this problem by
including a finer time unit in the filename. In general, always include
the time unit smaller than the file frequency time unit.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>