Update mvapich2 settings in GCHP example bashrc to give more control to SLURM
MPI implementation mvapich2 was incorrectly assigning core bindings when
greater than one node was used. The result was lack of performance
scalability with increasing number of nodes. Turning off usage of core
bindings with mvapich2 (MV2_ENABLE_AFFINITY=0) and use of a shared memory
pool when binding cores (MV2_USE_SHARED_MEM=0) corrects this issue. Please
note, however, that this puts SLURM in control of core bindings and the
available memory pools. The example SLURM bashrc file included in the
bashrcSamples subdirectory within the GCHP run directories should be
used for guidance on how to appropriately configure SLURM to do this.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>