Update Rn-Pb-Be simulation to use 72 levels by default
Then Rn-Pb-Be simulation previously used the reduced 47 levels by default.
To simplify comparison to GCHP and to better track the stratosphere, this
simulation now uses 72 levels by default. This requires passing the make
option NO_REDUCED=y if it's an RnPbBe simulation and also updating the
GEOS-Chem restart file to 72 levels.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>