Now create HISTORY.rc files from HISTORY.rc.template files
Made a modification to the perl/gcUnitTest, perl/gcCopyRunDirs,
and perl/gcCreateDiffTest scripts to create the HISTORY.rc files
in each GC run directory from a corresponding HISTORY.rc.template.
This will make sure that we can set the frequency setting to match
the end time in input.geos.

NOTE: In a following commit, we will change the HISTORY.rc files
to HISTORY.rc.template files with replacable tokens.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>