Rename runs/shared_inputs/HISTORY.rc files to HISTORY.rc.template
Because we need to create a HISTORY.rc file with the corresponding
frequency and duration as the start & end dates in input.geos,
we have renamed the runs/shared_inputs/HISTORY.rc files to
HISTORY.rc.template.

We have also added replaceable tokens {FREQUENCY} and {DURATION} that
will get populated with the proper format string when creating
HISTORY.rc files in the run directory.

Also, created a new folder runs/shared_inputs/RRTMG, which for the
moment just links to files in run/shared_inputs/standard.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>