Routine getDuration now accounts for minutes
Modified the &getDuration routine in UtUtils.pm to account for minutes
(e.g. if the run is less than an hour, to make sure that we will
populate the frequency and duration tags of the HISTORY.rc file
accordingly.

NOTE: May need some further testing but should be OK for now.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>