Simplify runs/shared_inputs directory
The unit tester was failing because the HEMCO_Config.template files
for the tropchem and standard simulations was not consistent with the
HEMCO_Config.template file for the benchmark simulation -- the same
HEMCO_Config file can be used for all three simulations. To avoid issues
and confusion, here we consolidated full-chemistry input files into a
single directory called shared_inputs/fullchem/ so that all
HEMCO_Config.template and HISTORY.rc.template files for full-chemistry
simulations are in one place, which makes updating those files easier.

Also updated links in run directories so that all FAST-JX input files
are linked directly to the shared_inputs/fast_jx directory instead of
having symbolic links that point to other symbolic links.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>