Add optional INIT_COMMANDs string to UnitTest.input
The INIT_COMMANDS string of UnitTest.input can be used to add Unix
commands (such as to source the bashrc or to load modules).  These
commands will be placed just underneath the SLURM tags and just above
the export OMP_NUM_THREADS declaration.

For example, depending on your computer cluster, it might be necessary
to source the bashrc and load modules from within the job script.
This can now be done with INIT_COMMANDS.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>