Create rst sym links in GCHP benchmark rundir for file with HEMCO rst vars
Restart symbolic links for the GCHP benchmark simulation restart files
previously linked to file that did not contains HEMCO restart variables.
Now there is also the option of using restart files with the HEMCO
restart variables. These should be used when doing the 1-month benchmark,
but not the 1-year benchmark. Also in this commit, benchmark restart
files are now included for four different resolutions: c24, c48, c90, and
c180.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>