Update HEMCO_Config.template files to use OMI-based volcanic emissions
The AeroCom volcanic SO2 emissions extension is used to read in and
process OMI-based volcanic emissions for 2005-2012 (provided by Cui Ge).
The OMI emissions use the same format as the AeroCom emissions. The
option "Volcano_Source : OMI" is specified by default for this extension
so that the closest available year will be used if the simulation year
is outside of 2005-2012.

The former VOLCANO base emission option is now obsolete and has been
removed from the HEMCO_Config.template files.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>