Set EXPID to ./GEOSChem for all GC-Classic HISTORY.rc.template files
We now use the EXPID parameter in HISTORY.rc to form the front of the
file names.  Set EXPID to "./GEOSChem" by default, which will write
files that begin with "GEOSChem" in the current run directory.
This will avoid breaking the Unit Tester.

Users who wish to send netCDF output to a sub-folder of the run
directory may e.g. set EXPID to "./OutputDir/GEOSChem", etc.

This now matches the behavior of GCHP.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>