Fix wrong link to HISTORY.rc.template for marinePOA simulations
The links to the HISTORY.rc.template files in the runs/4x5_marinePOA
and runs/2x25_marinePOA run directories were incorrect.  These were
symbolically linked to the HISTORY.rc.template.benchmark file in
runs/shared_inputs/fullchem.

But the marinePOA simulation is built off the standard simulation
(without the complex SOA) instead of the benchmark simulation,
and so this link is incorrect.  We have now corrected the symbolic
links in the runs/2x25_marinePOA and runs/4x5_marinePOA run
directories.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>