Comment out CEDS emissions for CO2 and CH4 in all fullchem HEMCO_Config.rc
CO2 is not an advected/emitted species in the full-chemistry simulations.
The CO2 CEDS emissions are in the HEMCO_Config.rc file but are now
commented out. Users may easily enable them if CO2 is added as an
advected species.

CH4 is now an advected species in all full-chemistry simulations, but
the surface concentrations are prescribed at the start of a simulation
using monthly CH4 distributions from Lee Murray for v11-02 and later and
using concentrations defined by latitude bands prior to that. CH4 emissions
from CEDS should not be used. They are still in the file, however, but
commented out. Users may enable them if they wish by making the
appropriate updates.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>