Revert to NH3_geos.4x5.nc file to avoid excessive NH3 emissions
The NH3_geos.1x1.nc file (which Eleanor Morris of York U.) had
recommended that we use) is problematic.  We believe this file has
a regridding error that leads to an excessive amount of NH3 natural
emissions (20x more than we would expect).

The quick fix is to revert to the NH3_geos.4x5 file, which is what
we had been using all along.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>