Change biofuel file resolution in GCHP from 1x1 to 4x5 to match GC classic
There is a discrepancy between the 1x1 and 4x5 biofuel files according
to the ExtData file README. The discrepancy is up to a factor of two
difference in SO2 in certain regions. The README states that we therefore
recommending use the 4x5 and not the 1x1 file. The 4x5 file was already
listed in the GCHP HEMCO_Config.rc but the filename was superceded by
the 1x1 file listed in ExtData.rc.

This update should bring greater parity between GCHP and GEOS-Chen Classic.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>