Turn on CEDS inventory for GCHP fullchem simulations
This update accompanies a file update to the CEDS inventory required
for compatibility with MAPL in GCHP. An extraneous bounds dimension and
variables that use it were created during pre-processing CEDS data. Their
presence in the file causes MAPL to throw a dimension read error during
initial read. Remove the dimension and variables solves the issue.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>