Minor updates for perl/Makefile and perl/difftest/Makefile
perl/Makefile:
(a) Set FPEX=n if it is not explicitly defined.

perl/diffteset/Makefile:
(a) Set FPEX=n if it is not explicitly defined.
(b) Set NO_REDUCED=n if the simulation is RnPbBe (it uses the full grid).
(c) Add NO_REDUCED=$(NO_REDUCED) to the following targets:
     refonly, devonly, devlib, devrun

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>