Rename "runs/4x5_benchmark/v11-02.run" script to "benchmark.run"
We have renamed the v11-02.run script in the runs/4x5_benchmark folder
to "benchmark.run".  This should make the script less tied to a particular
version.

Added some comments as to how optional settings in the script can be
disabled.  Also put a reminder to change the id tag for each version.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>