Update HEMCO_Config.template* and ExtData.rc files for GFED4 and AFCID
The last year of GFED4 emissions is now 2016 (formerly was 2014).
We have therefore updated the GFED4 entries of all relevant
HEMCO_Config.template* files in the runs/shared_inputs/* folders.

For the AFCID (anthro PM2.5 dust) emissions, we now use container names

  DEAD_PM25FINE_1
  DEAD_PM25FINE_2
  DEAD_PM25FINE_3

because GCHP will hang in file I/O if all of the container names are
the same.  Also updated comments to denote that the AFCID emissions
will go into GEOS-Chem 12.1.0.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>