Explicitly turn off GEIA NH3 anthro and biofuel if CEDS or EDGARv43 is on
We discovered that HEMCO's default behavior is to ignore the GEIA NH3
anthro and biofuel emissions if CEDS is turned on.  This is done
"under the hood" and is echoed to the HEMCO.log file when VERBOSE=3.

In order to avoid relying on default behavior, we have explicitly
modified all relevant HEMCO_Config.template* files so that the
GEIA_NH3_ANTH and GEIA_NH3_BIOF containers are disabled if either
CEDS or EDGARv43 are turned on.

This update should not result in any differences in GC output, but
just makes the default HEMCO behavior more explicit.