Add script to archive GCHP run dir contents; rm gcchem* restart with make
Use shell script archiveRun.sh to archive output diagnostics, restarts,
config files, run script, env file, and logs to a subdirectory. This
allows clean reuse of a GCHP run directory without losing information
for a past run. The script is interactive, prompting the user to
name the subdirectory at the prompt. The archive subdirectory does not
need to exist prior to archiving.

Also in this commit, update the 'make clean_output' target in the
Makefile to delete gcchem* restart files (checkpoint files) when
cleaning the run directory. The restart files are automatically
archived when executing archiveRun.sh so the user can save them this
way if they do not want to lose them.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>