Remove HEMCO restart file from 4x5_benchmark run directory
The HEMCO restart file in 4x5_benchmark was dated 20130701, but we now
run 1-month benchmarks for 20160701. This meant that the HEMCO restart file
was ignored by HEMCO. Christoph Keller recommends only using a HEMCO
restart file for the date and time indicated in the filename. HEMCO will
automatically recognize when a restart file is not available for the date
and time required, and in that case HEMCO will use default values to
initialize those fields.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>