Update GCHP monthly diagnostic option to be set in the multirun script
There are performance trade-offs associated with using the GCHP monthly
diagnostic option in the run directory. The usage of multiple consecutive
jobs comes with the drawback of an additional initial file read for each
run. There is also potential for more wait times in the cluster queue.
The option is therefore removed from runConfig.sh and is set within
the multirun run script for advanced users only. After more testing is
done and the performance trade-off is better understood we may make the
option more visible.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>