Add improved error check for CODE_DIR in perl/Makefile
We now have added the correct algorithm for checking the validity
of the CODE_DIR in the perl/Makefile.  This will be the makefile
that gets copied to each GEOS-Chem run directory.

If CODE_DIR points to an missing source code directory, the compilation
process will halt with an error message.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>