Bug fix in time string in HEMCO_Config.template.GCHP
In the offline sea-salt, dust, and lightning NO fields, the time strings
in the HEMCO_Config.template.GCHP file (which goes into the folder
gchp_standard) were given as:

  2013-2016/1-12/1-31/0*

but this caused a Fortran error because the string "0*" was being
converted to an integer value, which caused a string-parsing error.
We have now changed the "0*" characters to just "0", which gets
around the error.

NOTE: The time frequency is handled in ExtData for GCHP.  But these
time strings in HEMCO_Config are read in and later ignored, so they
are susceptible to syntax errors.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>