Turn on CEDS ship emissions
CEDS ship emissions will now be used as the default global ship emissions
inventory. This inventory now overwrites ARCTAS SO2 and ICOADS CO and NO
emissions. The EMEP_SHIP entries have been updates to use a hierarchy of
10 so that they overwrite CEDS_SHIP emissions. In addition, a typo is fixed
here so that we use extension #102 for CEDS_SHIP when PARANOX is turned
on.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>