Update GCHP compile.log creation rules in the Makefile
A new GCHP compile.log file is created, overwriting the old, if compiling
with 'make compile_clean'. Otherwise the existing compile.log file is
appended to. This enables keeping the original log of compiling ESMF,
MAPL, and fvdycore, while also keeping a recording of compiling the main
components of GEOS-Chem.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>