Bug fix: GCHP will use multi-year CEDS files instead of yearly files
In ExtData.rc, the token %y4 was used in the file paths for the yearly CEDS
files. However, MAPL replaces the token with the current simulation year.
This causes issues if the simulation year is beyond the last year of CEDS
emissions available (2014). To avoid this issue for now, we will revert
to using the multi-year CEDS files in CEDS/v2018-04/ instead of the yearly
files in CEDS/v2018-08/.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>