Clean up mask file entries and make sure masks are on global grids
(1) Lizzie updated the Africa mask file to put the mask on a global grid.
    Here we also rename AF_MASK to AFRICA_MASK for clarity.

(2) Update the Canada, Mexico, and CONUS mask files (and their mirror masks
    files) used in the CH4 simulations to put the masks on a global grid.
    Also fixed the CONUS mask (and mirror) so that it is shifted north 1
    degree to prevent a gap between the Canada and CONUS masks.

(3) Now point to mask files in MASKS/v2018-09/. These are the same files
    that were in MASKS/v2014-07/, but the directory has been greatly
    cleaned up so that it doesn't include masks that aren't currently used
    in GEOS-Chem.

(4) Remove entries for old NA and USA masks that are not used. Also remove
    lines for the obsolete Cooke BC/OC inventory.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>