Revert to regional mask files for CH4 simulations
In an earlier commit, the Canada, Mexico, and US mask files used in the
CH4 simulations were updated so that they were on a global grid. This was
done for GCHP, which doesn't seem to work with regional mask files. However,
using global 0.01x0.01 causes a drastic increase in run time and memory
usage for the global CH4 simulations. To avoid this, we will revert to
the regional mask files for now.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>