Update HEMCO_Config.template for CO2 sim to avoid double-counting
Beata Bukosa and Christoph Keller have modified the HEMCO_Config.rc file
to add emissions into the tagged CO2 species via HEMCO instead of
manually adding them in co2_mod.F.  This fixes a problem where the CO2
emissions were being double-counted.

In addition to this update to the HEMCO_Config.template file, the code
in co2_mod.F (in the GEOS-Chem repository) that manually added emissions
into the species array has now been disabled.

P.S. Also updated the Makefile version number to 12.0.2, which is the
version in which this update will be included.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>