Add Restart collection to the HISTORY.rc.template files
The Restart collection will eventually replace the GEOS-Chem restart files.
For now, we hardcode the frequency and duration so that the restart file
fields are saved out daily to a monthly file. Several fields from State_Chm
and State_Met are saved out to the Restart collection in addition to the
instantaneous species concentrations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>