Remove species fluxes lines from all input.geos.template file
The "Species fluxes" lines in the input.geos.template files were used
to populate the Input_Opt%LSPECRADMENU field.  But we now get the
RRTMG requested flux outputs from HISTORY.rc, so these lines are no
longer needed.

For the time being, we force-feed the Input_Opt%LSPECRADMENU variable
from the flux outputs selected in HISTORY.rc.  We need to keep
Input_Opt%LSPECRADMENU as long as the bpch diagnostics are present,
because this variable is used in diag3.F.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>