Add common set of tracers between GEOS and GEOS-Chem
The transport Working Group has proposed implementing additional tracers
in GEOS-Chem for benchmarking purposes and diagnosing transport in
GEOS-Chem compared to GEOS-5. The conversation is documented on the wiki
at:

  http://wiki.seas.harvard.edu/geos-chem/index.php/Transport_Working_Group#Common_set_of_tracers_in_GEOS-Chem_and_GEOS

Here we add several of those species. Please note that validation is ongoing
and that data files specified in HEMCO_Config.rc for the Rn-Pb-Be simulation
have not been copied to the shared data directories yet.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>