Update US land mask applied to NEI2011 emissions
The old land mask  (NEI11_MASK) applied to NEI2011 emissions covered regions
of northern Mexico and southern Canada, acting to zero out regional emissions
there. To avoid this, we use a new land mask for the US that was created from
a high-resolution shapefile of country borders and saved out on a 0.01x0.01
grid.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>