Use 0.1x0.1 mask file for US instead of 0.01x0.01 file
The high-resolution 0.01x0.01 CONUS mask file required a TileFile for GCHP
that was too large for tempest to create. We will use a coarser mask file
in both GCHP and GCC to get around that issue. This update passes a difference
test for the geosfp_4x5_benchmark simulation and will not impact the 1-month
benchmark.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>