Remove NC_DIAG option from Makefiles
The netCDF diagnostics were originally implemented as an alternative
to, and eventual replacement of, the BPCH diagnostics. A NC_DIAG CPP
switch was put in place to separate the BPCH and netCDF diagnostic
code. However, now NC_DIAG should always be used to enable saving out
the GEOS-Chem restart file (via the Restart collection in HISTORY.rc).
BPCH diagnostics can still be used in parallel with netCDF diagnostics
until they are fully validated and retired. BPCH diagnostics can be
turned on/off in input.geos and netCDF diagnostics can be turned on/
off in HISTORY.rc.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>