Update test for NO_CODE_FOUND in runs/shared_inputs/Makefiles/Makefile
The test:
NO_CODE_FOUND:=$(shell test -d $(CODE_DIR) -a ! -h $(CODE_DIR); echo $$?)

checks if CODE_DIR is a directory and not a symbolic link. But now that
we create the ./CodeDir symbolic link to the code directory, this test
would fail and cause the build sequence to halt with an error.

The simple fix is to remove the "-a ! -h $(CODE_DIR)", which is the
logical condition "not a symbolic link".  Removing this causes the
build sequence to proceed properly.

Also if ./CodeDir is a broken link, the run will halt with an error.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>