Update MODIS_XLAI entries for proper interpolation
In the MODIS_XLAI section of the HEMCO_Config.template files,
we have changed the time entries from:

   2005-2011/1-12/15/0 C

to:

   2005-2011/1-12/1-31/0 I

This will make HEMCO re-compute the interpolation factors at
the start of each new day.  Also note: a corresponding update
to HEMCO version 2.1.012 in the GEOS-Chem source code must
accompany this commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>