Use relative paths in Ref symbolic links to allow moving DiffTest directory
Previously the symbolic links in the Ref directory which pointed to files
in the Dev directory used full paths to the Dev directory. Therefore the
symbolic links broke if the entire DiffTest directory was moved, since
the Dev directory location changed. Using ../Dev/filename in Ref symlinks
corrects this issue.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>