Specify always use simulation year for restart and met fields in HEMCO_Config.tc
A user brought to our attention an issue caused by setting the emissions
year in HEMCO_Config.rc when using GEOS-Chem 12.1.0. Starting in that
version, we read in met fields and the GEOS-Chem restart file via HEMCO.
Setting the emissions year seems to also override the year for those
files when the user probably still wants to use the simulation year.

To fix this issue, we have added the "Y" option to the time cycle flag to
tell HEMCO to always use simulation year for certain fields instead of the
emissions year.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>