Update permissions in copy_1mon script for benchmarks
We now include benchmark plots created in GCPy in the output directory
of the 1-month benchmark run directories. The section where permissions
are changed in copy_1mon is now updated to account for the new directories
and files.

We also now remove CodeDir. Starting in this version, we create a symbolic
link to CodeDir when creating run directories. There is no need to store
the entire code directory permanently. The information about the code version
and the specific commit used can be found in the lastbuild.mp file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>