Add JNoonFrac to the JValuesLocalNoon collection
We have added the JNoonFrac field to the JValuesLocalNoon collection
for all of the HISTORY.rc.template* files pertaining to full-chemistry
simulations.

To obtain the proper noontime J-values, one must divide the J-values
(which are now accumulated values) by JNoonFrac (which, when time
averaged, is the fraction of the time it was local noon in each
grid box).

We have also updated the version number from 12.3.0 to 12.3.1 in
the default run directory Makefile.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>