Turn off ND06 dust emissions diagnostic in bpch output
When the ND06 dust emissions diagnostic is turned on, the simulation will
crash because the diagnostic cannot be found in HEMCO. This is because
the diagnostic is created for the DustDead extension in hcoi_gc_diagn_mod.F90.
However, in 12.4.0 we turn off the DustDead extension and use offline
dust emissions instead. We could resolve the error by changing the ExtNr
for this diagnostic to 0 (base emissions) in hcoi_gc_diagn_mod.F, but then
if users decide to turn on the dust extension they will have to go back
in and change it again. To simplify matters, we will just turn off ND06
in the bpch output. The bpch diagnostics will be retired soon anyway and
we now recommend users get the emissions diagnostic output from the
HEMCO diagnostics file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>