Replace hardcoded resolution strings for offline emissions with tokens
HEMCO will automatically choose which resolution offline emissions to use
based on the meteorology field for a given simulation. To allow for this
flexibility the hardcoded resolution string "0.25x0.3125" has been replaced
with {NATIVE_RES} and "025" has been replaced with {LATRES}. The correct
resolution string will automatically be added when users create a run
directory.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>