Remove Met_AREAM2 from HISTORY.rc files
The field Area_M2 has been moved from State_Met to State_Grid in this version,
so we have removed it from the StateMet collection in HISTORY.rc. There is
no need to output the Area_M2 field in a collection because this field is
automatically included in all netCDF files produced by the History component.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>