Rename nested CH to AS for GEOS-FP
For simplicity, the nested "CH" grid for GEOS-FP will actually read in
the AS met fields. These are the same met fields, but on a larger domain
to allow for more flexibility when defining custom grids. This also
simplifies differences between MERRA-2 and GEOS-FP simulations because
MERRA-2 uses the AS domain and not the CH domain. The domain extent for
the nested GEOS-FP simulation remains unchanged here -- that is defined in
input.geos. In that file, users can either choose to use the smaller
GEOS-FP domain previously referred to as "CH", the larger domain consistent
with MERRA-2 AS, or their own custom domain by changing the lat/lon bounds.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>